move into `Redirects` directory
