+++ routes/cp.php
@@ -24,7 +24,7 @@ Route::prefix('seo-pro')->name('seo-pro.')->group(function () {
     Route::patch('section-defaults/taxonomies/{seo_pro_taxonomy}', [Controllers\CP\TaxonomyDefaultsController::class, 'update'])->name('section-defaults.taxonomies.update');
 
     Route::resource('errors', Controllers\CP\ErrorController::class)->only('index');
-    Route::resource('redirects', Controllers\CP\RedirectController::class)->except('show');
+    Route::resource('redirects', Controllers\CP\Redirects\RedirectController::class)->except('show');
 
     Route::post('preview', Controllers\CP\PreviewController::class)->name('preview');
 });
